Add unit tests for values.
