+++ cocoasm/statement.py
@@ -242,7 +242,7 @@ class Statement(object):
             self.size = self.instruction_bundle.get_size()
             return
 
-        if self.operand.value.requires_resolution():
+        if not self.operand.value.resolved:
             self.operand.resolve_symbols(symbol_table)
 
         if self.instruction.is_short_branch or self.instruction.is_long_branch:
